PLST-6353:[PATCH] Process Put Aways Get Next Tote (#1414)
Adjusted funcitonality for get next tote, to be based on the subscribe event instead of the function call. Just using the function caall casued the tote id to get out of synch with other screens and resulted in duplicate totes across batches. Adjust anywhere we incremented the tote id and putnthat inot a subscribe event.