PLST-6525:[PATCH] Resolved issues with Move Screen (#1440)
Resolved issue with slected Move From Location appearing on the Move To Tab. Resolved issue with selecting a Move To location triggering a redraw and losing focus on the selcted location. Resolved bug where the wrong value was being referenced from the selected rows in order to get the inv map ids to pass back to the API