PLST-10220:[PATCH] LogixPro | Essity | Field name mapping is not working on multiple fields and dropdowns (#1516)
* Refactor column filter to use tableColumns directly

Updated column-filter-component to use tableColumns instead of filteredTableColumns and removed the filteredTableColumns getter. Adjusted routeid-list to pass a filtered searchDisplayedColumns prop to the column filter component for improved consistency and type safety.

* Add replacePlaceholder attribute to mat-option

Added the replacePlaceholder attribute to each mat-option in the column filter component to support new functionality or customization for options.