PLST-10925:[PATCH] LogixPro | v15 | Transaction counts mismatch between Inventory Details tab and Transaction grids (#1572)
* Trigger data refresh on item number selection

Added calls to getContentData() after setting item number filters in open-transaction-on-hold, reprocess-transaction, and transaction-history-list components. This ensures the displayed data updates immediately when an item number is selected. Also introduced a 500ms delay before processing item number changes in transaction-history-list to improve UX.

* fixation of transaction history filter issue

when try to coming from inventory page through item number the data is not reflecting properly as it is making the unnecessary api calls and over ride the data.

* reset logix fixation and assign type to relevant variables and functions

---------

Co-authored-by: HafizShehbaz <shehbaz.hasan@enterprise64.com>