PLST-11505:[PATCH]  TH - UAT Issue - Induction Manager - Process Put Away - Duplicate Tote ID (#1602)
* Prevent duplicate Tote IDs in put-aways table

Added a check in updateToteID to prevent entering a Tote ID that already exists in the table. Displays an error message and clears the input if a duplicate is detected.

* Refactor duplicate Tote ID error message handling

Moved the duplicate Tote ID error message to the ToasterMessages constants and updated its usage in process-put-aways.component.ts for consistency and maintainability.