PLST-11073:[PATCH] TH - UAT Issue - LogixPro | Admin | Inventory | Updating "Supplier Item ID" on Details tab also updates "Supplier ID" on others tab of Inventory screen (#1618)
* Update supplierName form control and input binding

The supplierName field in the inventory master form now initializes from getInvMasterData and enforces a max length of 50. The corresponding input in others.component.html is now bound to supplierName instead of supplierItemID, with context menu updated accordingly.

* Rename supplierNo to supplierName in admin component

Updated the form control and related references from 'supplierNo' to 'supplierName' in both the template and TypeScript file. Also changed the label from 'Supplier Number' to 'Supplier ID' for clarity.