PLST-12149:[PATCH] Inventory Map Add Loc | Comment out item control v… (#1643)
* PLST-12149:[PATCH] Inventory Map Add Loc | Comment out item control value reset in performClear

The line resetting the item control value in performClear was commented out, possibly to prevent unintended clearing of the field when no data is returned. Other field clearing logic remains unchanged.

* PLST-12149:[PATCH] Inventory Map Add Loc | Remove commented code from performClear method

Deleted an unnecessary commented-out line in the else block of the performClear() method to improve code clarity.

* Revert "PLST-12149:[PATCH] Inventory Map Add Loc | Remove commented code from performClear method"

This reverts commit 8206a7fdf6a4a7ec7724feef8534f692f71bbd93.

* PLST-12149:[PATCH] Inventory Map Add Loc | Fix item list clearing logic in map location dialog

Refactored the else block to ensure itemNumberList is reset and fields are cleared when no data is returned. Removed commented-out code for better clarity.