PLST-5878:[FEATURE] SSP - LogixPro - Process put-aways doesn't allow you to complete blind put away without overreciept set to allow. (#1651)
* Add 'New Put Away for Same SKU' preference and validation

Introduces a new admin preference 'New Put Away for Same SKU' with UI toggle and form control. Updates related components to respect this preference, disables the button when not enabled, and adds improved validation and error messaging for quantity assignment in tote extension dialogs.

* Disable 'New Put Away For Same SKU' button based on preference

The 'New Put Away For Same SKU' button is now disabled if 'data.imPreference?.newPutAwayForSameSKU' is false or undefined. Also fixed a bug in the TypeScript file where the wrong OTID was being validated in 'validateOverReciept'.